Allow the `gen_requirements` to be run by anyone.
We were previously including the private registry as extra index, but
this would cause problems for users who don't have access to this. By
not including this make it such that our project can't have anything
from the private package (which is desirable in this context)

This commit also fixes an issue with the `pip-compile` version not
including a command line option for the resolvers. Updating the version
resolves this.

This commit also includes pandas in the dev requirements which allows
the tests to pass.