TECH-661. Add a concurrency limit on the async session
We were previously allowing the session to have an unbounded number of
requests in flight at given time. This was causing issues for us.

This commit adds a new parameter `max_connections_limit` to the
`get_async_session` function. This parameter if passed will limit the
connection pool size of the underlying httpx session. See:
https://www.python-httpx.org/advanced/#pool-limit-configuration

The number of connections 200 was selected arbitrarily.