TECH-661. Limit the number of background tasks at a given time.
We experienced issues with this script with too many requests in flight.
Instead of having everything in the queue and waiting, we here limit the
number in flight by inspecting the `background_tasks` set. The task will
sleep if there is more than a limit and then once the background tasks
queue decreases we continue to get work off of the queue.