Make the fake proxies more realistic
When we added this example recently we added it with the same string for
both the http and https proxies. It's more likely that the https proxy
will have an `https` prefix, and that the port will be a conventional
port for the protocol (as in "443" is a more common https protocol and
"8080" is a more common "http" protocol).